requirements.txt
requests>=2.31.0
tenacity>=8.2.2
pydantic>=2.5.0
pydantic-settings>=2.2.1
pandas>=2.2.0
openpyxl>=3.1.2
streamlit>=1.32.0
faiss-cpu>=1.7.4; platform_system!='Windows'
openai>=1.3.5
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.0
